+++ modules/os_functions/builtin/cpp/systemBuiltin.cpp
@@ -58,7 +58,7 @@ Nelson::OsFunctionsGateway::systemBuiltin(Evaluator* eval, int nLhs, const Array
         Error(ERROR_WRONG_ARGUMENT_1_TYPE_STRING_EXPECTED);
     }
     int ierr = 0;
-    ArrayOf ret = SystemCommand(cmd, ierr);
+    ArrayOf ret = SystemCommand(cmd, ierr, eval->haveEventsLoop());
     if (bEcho) {
         Interface* io = eval->getInterface();
         std::wstring msg = ret.getContentAsWideString();