+++ modules/os_functions/src/include/SystemCommand.hpp
@@ -24,8 +24,6 @@
 //=============================================================================
 namespace Nelson {
 NLSOS_FUNCTIONS_IMPEXP ArrayOf
-SystemCommand(const std::wstring& command, int& ierr);
-NLSOS_FUNCTIONS_IMPEXP std::wstring
-SystemCommandW(const std::wstring& command, int& ierr);
+SystemCommand(const std::wstring& command, int& ierr, bool withLoopEvents);
 } // namespace Nelson
 //=============================================================================